[FEATURE] First version of YKUSH handler
This commit adds the first version of the YKUSH handler to the usb monitor. So
far, the application can detect when hubs or devices are added/removed. If we
discover more hubs with power-control support, a lot of the features we have
added are generic.

Next step is to add support for switching on and off.