[FEATURE] Finish support for YKUSH hub
This commit finishes, at least for now, adding support for the YKUSH hub to the
usb monitor. After this commit, devices connected to YKUSH hub are pinged every
five seconds. If we loose five packets in a row, then a device is power-cycled.
A device (or rather power) is reset if we see no device for twenty seconds.

The commit also adds some core architecture features:

* The whole application, for now, is driven by the libusb blocking wait
  function.
* Every USB call is changed to be asynchrnous, so we never block.
* Handling the async calls are done in two ways:
    - When waiting to send a request, we use our own timeout queue.
    - While waiting for a reply to a request, we use the timeout in the libusb
      transmit-call.