[FEATURE] Handle hub configuration failures properly
If we fail to configure a hub, say device can't be claimed, then we currently
never retry. The same applies if we don't detect a devie when it is connected.
To add support for detecting failed hubs, we now ever 30 sec check the device
list for an devices that are missed. It could be that it is a bit of a waste to
do this every 30 sec, but with the number of devices we have now it does not
impact performance.

Also, standard list-loops have been rewritten to use for-each.