[FEATURE] Support switching USB ports using GPIO
This commit adds a GPIO handler. GPIO works slightly different to other devices
in that we rely on our timeout for switching the device back on (there is no
USB-message for switching on/off). Otherwise, the logic is the same as for
YKUSH.

This commit also makes a different change. The ctx pointer was moved from hub to
port, as not all ports have a parent hub that we monitor (for example GPIO).