[FEATURE] Support logging to file
This commit adds basic logging to file. File is specified as a command-line
argument, and if no file is specified then stderr is used. Each line contains a
timestamp and the message. A crude rate-limiting is used on the ping messages,
to prevent them from flooding the log-file.

I assume that a different tool will handle log rotation.