[FEATURE] Add a proper event loop
This commit adds a proper event loop, which is a neccessary step for how we plan
to add support for user input. The event loops provides all the basic features,
as well as support for a proper timer queue. Our different timeout events should
also be converted to use this queue.

Queue is currently a list which is not a problem for now, but could lead to
scalabiity issues later.