[FEATURE] Add REST API for switching on/off ports
This commit adds support for a small REST API that can be used to get state of
ports with devices connected, as well as switch on/off ports. I chose not to
export ports without any device connected, as those ports are automatically
restarted every minute by usb monitor.

We currently only support one GET operation and one POST operation. This means
that the path does not matter, we only look at HTTP method.

GET is used to get the status, vid and pid of the ports and an example output
is:
{"ports":[{"path":"3-1-2-5-4-3","mode":1,"vid":4817,"pid":5382}]}

In order to restart one or more devices, a POST request must be sent. The syntax
is as follows:
{"ports": [{"path":"3-1-2-5-4-3", "cmd": 1}]}

path is the bus number + path to device. cmd is currently ignored, but in the a
future version we will support only switching on or off devices.

Note that there is lots of room for improvement in the "server". There is for
example currently no timeout, so an application can be hanging forever. However,
only root can read/write to the socket, so it is fair to assume that these will
be well-behaved applications. Still, we have a timer infrastructure, so it
should be fixed.

Also, support for generic hubs has been temporarily disabled. Some more work is
needed before they are stable, particularly with cascading devices.