[FIX] High CPU usage
We need to prove the cauase of the 100% CPU usage bug. The main culprit seems to
be misbehaving hubs/devices. When this happens, I see the following in the log
files:

[08:43:53 05/07/2015]: Device: 0835:8500 removed
[08:43:53 05/07/2015]: Type: GPIO Path: 1-1 State: 1 Pwr: 1  Device: 19d2:1403

[08:43:54 05/07/2015]: Device: 0835:8500 added
[08:43:54 05/07/2015]: Type: GPIO Path: 2-1 State: 1 Pwr: 0  Device: 0835:8500
[08:43:54 05/07/2015]: Type: GPIO Path: 1-1 State: 1 Pwr: 1  Device: 19d2:1403

USB ping has not failed (then we would see Ping failed) and hub is not
disconnected, so only cause is that there is some internal crash in hub. Device
is added again right afterwards, but note that power is 0. Power is only updated
when we reset port, so most likely the reset all ports function has executed.

When device is added again (event is cached or something similar), we hit
undefined behavior and 100% CPU. Solve this by checking if port state is reset
when we get the device added again. Since port will be restarted, there is no
point adding device or add any caching. State is moved back to IDLE after reset,
so then device will be added as normal.