[FEATURE] Support enabling/disabling USB ports
This commit adds support for enabling/disabling (so just restarting) USB ports.
In order to disable/enable a port, pass command 2/1 in the POST JSON request.
Disable/enable is currently not stored anywhere, so it will (currently) be lost
when usb monitor is restarted. Also, all ports are assumed to be enabled when
usb monitor is started. This is fine, any disabled port will be enabled by the
flip/flop that is run every 30 sec.

The implementation is mostly straight forward and the only slightly complicated
part is the YKUSH handling, since we send async requests to the device. I have
tested and make sure we support every possible combination of
PING/RESTART/ENABLE/DISABLE I could think of, but there is probably some that I
have forgotten.