Implement first version of port probing
A hardware vendor we are working with switched which USB port is
controlled by which GPIO, and there is no easy way to check the
port/gpio mapping of a device. In order to work around this issue, we
need to probe each GPIO and detect which device gets connected. The
current approach is as follows (on devices where we do probing):

* First, we disable all GPIO that controls USB ports.
* We wait five seconds for the ports to be switched off and any devices
to disconnect. Five seconds seems to be more than enough, based on our
testing with the affected device.
* After the timeout has expired, we turn on the first port we find in
our list. We give the device in the port 60 seconds to connect. 60
seconds is enough (and then some) for all our modems to boot + connect,
at least it has been during testing.
* When device is connected, we check if the path matches what is stored
for the port we enabled. If not, we swap the USB device path between the
port that matched and the port that was enabled.
* If we match, or timeout expires, we mark that probing is done and
enable the next port.
* In addition, for timeout we disable the enabled port. The reason for
doing this is to prevent devices from connecting after we are done
probing. We then wait five seconds before enabling the port again,
blocking any connection attempt.
* When we have probed all devices, we write the new configuration to
file and resume normal operartion.
* Next time usb-monitor is started, it will read the configuration from
file.

Limitations:
* Probing works best when a device in inserted in all ports that will be
probed.
* We only support swapping. This means that we can't discover new paths,
only incorrect mappings between pins and paths.
* Assumes that usb-monitor has full control over usb/pins.

I have tested this feature by first making sure that it work, and then
trying to trigger the different error cases + ensuring that all
different global timers are triggered. Feature seems to work fine, I
have not been able to break probing and we correctly fix the mapping on
"broken" devices.