gpio: Write port map to shadow file
We should write the port map to a shadow file, and then do link when
file is written. This ensures that we don't end up with broken
configurations. One cases where we have seen broken configs, is when
router is rebooted after file is created and before write is flushed to
disk.

Link is atomic, so it is either successful or not. This means that when
link is successful, then we have a known good config.

This reverts commit c0d12a567b2510c8a5a111a54cb389d70829dab0.