Speed up bad ID check
Do the bad ID check after a device has been connected, instead of
waiting 30 seconds for a timer. Speeding up bad id does not disturb
probing or the normal port restart. We do not check bad id when a port
is probed, and the normal port restart does not restart ports that are
being restarted (puh).